# backend/src/main/java/com/c2certi/tms/auth/api/LoginRequest.java
package com.c2certi.tms.auth.api;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LoginRequest(
    @NotBlank(message = "Username is required.")
        @Size(max = 50, message = "Username must be at most 50 characters.")
        String username,
    @NotBlank(message = "Password is required.")
        @Size(max = 128, message = "Password must be at most 128 characters.")
        String password) {

  @Override
  public String toString() {
    return "LoginRequest{username=" + username + "}";
  }
}