# specs/001-support-ticket-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Ticketing Management System API
  version: 1.1.0
  description: |
    REST API for support tickets. All endpoints except `POST /auth/login` require a valid
    `TMS_SESSION` cookie (HttpOnly signed JWT). State-changing requests require the
    `X-XSRF-TOKEN` header matching the `XSRF-TOKEN` cookie. Errors use RFC 9457
    `application/problem+json` (see `Problem`).
servers:
  - url: /api/v1
security:
  - sessionCookie: []
tags:
  - name: auth
  - name: users
  - name: tickets
  - name: comments

paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Sign in and receive session cookie
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Signed in. Sets `TMS_SESSION` cookie.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401':
          description: INVALID_CREDENTIALS (generic message)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  /auth/logout:
    post:
      tags: [auth]
      summary: Sign out (clears session cookie)
      responses:
        '204': { description: Signed out }

  /auth/me:
    get:
      tags: [auth]
      summary: Current signed-in user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthenticated' }

  /users:
    get:
      tags: [users]
      summary: List active users (assignee picker)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserSummary' }
        '401': { $ref: '#/components/responses/Unauthenticated' }

  /tickets:
    get:
      tags: [tickets]
      summary: List/search/filter tickets (paginated, newest-updated first)
      parameters:
        - name: q
          in: query
          description: Case-insensitive keyword over key, title, description (literal match)
          schema: { type: string, maxLength: 100 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/TicketStatus' }
        - name: view
          in: query
          description: |
            `mine` = reported by current user (default), `all` = every ticket,
            `assigned` = assigned to current user with status OPEN or IN_PROGRESS, ordered by
            priority (CRITICAL first), then oldest created, then reference. `q` and `status`
            combine with AND, so `view=assigned&status=RESOLVED|CLOSED` returns no items.
          schema: { type: string, enum: [mine, all, assigned], default: mine }
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TicketPage' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [tickets]
      summary: Create ticket (status OPEN, reporter = current user)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTicketRequest' }
      responses:
        '201':
          description: Created
          headers:
            Location:
              description: /api/v1/tickets/{ticketKey}
              schema: { type: string }
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TicketDetail' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }

  /tickets/summary:
    get:
      tags: [tickets]
      summary: Counters for the signed-in user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TicketSummaryCounts' }
        '401': { $ref: '#/components/responses/Unauthenticated' }

  /tickets/{ticketKey}:
    parameters:
      - $ref: '#/components/parameters/TicketKey'
    get:
      tags: [tickets]
      summary: Ticket details with comments and history
      responses:
        '200':
          description: OK
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TicketDetail' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [tickets]
      summary: Update title, description, priority and/or assignee (not status)
      description: |
        Only supplied fields are changed. Unknown properties (including `status`) are rejected
        with 400. Rejected with 409 TICKET_CLOSED when the ticket is CLOSED.
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTicketRequest' }
      responses:
        '200':
          description: Updated
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TicketDetail' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/VersionConflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }

  /tickets/{ticketKey}/transitions:
    parameters:
      - $ref: '#/components/parameters/TicketKey'
    post:
      tags: [tickets]
      summary: Change status (only OPEN→IN_PROGRESS→RESOLVED→CLOSED)
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransitionRequest' }
      responses:
        '200':
          description: Transition applied
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TicketDetail' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: INVALID_STATUS_TRANSITION (ticket unchanged)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              example:
                type: about:blank
                title: Status change not allowed
                status: 409
                detail: A ticket in OPEN can only move to IN_PROGRESS.
                code: INVALID_STATUS_TRANSITION
                currentStatus: OPEN
                allowedNextStatus: IN_PROGRESS
                traceId: 4bf92f3577b34da6
        '412': { $ref: '#/components/responses/VersionConflict' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }

  /tickets/{ticketKey}/comments:
    parameters:
      - $ref: '#/components/parameters/TicketKey'
    get:
      tags: [comments]
      summary: Comments, oldest first
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Comment' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [comments]
      summary: Add comment (author = current user); 409 TICKET_CLOSED if closed
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCommentRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: TMS_SESSION

  headers:
    ETag:
      description: Quoted ticket version, e.g. `"3"`
      schema: { type: string }

  parameters:
    TicketKey:
      name: ticketKey
      in: path
      required: true
      schema: { type: string, pattern: '^TMS-[1-9][0-9]*$' }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag from the last read of the ticket
      schema: { type: string }

  responses:
    ValidationFailed:
      description: VALIDATION_FAILED with per-field errors
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: about:blank
            title: Please correct the highlighted fields
            status: 400
            detail: 2 fields need attention.
            code: VALIDATION_FAILED
            errors:
              - { field: title, message: Title must be between 3 and 150 characters. }
              - { field: assigneeId, message: Please choose an assignee. }
            traceId: 4bf92f3577b34da6
    Unauthenticated:
      description: UNAUTHENTICATED (missing/expired session)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: TICKET_NOT_FOUND
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: TICKET_CLOSED (closed tickets are read-only)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    VersionConflict:
      description: VERSION_CONFLICT (ticket changed since last read; reload)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    PreconditionRequired:
      description: PRECONDITION_REQUIRED (If-Match header missing)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }

  schemas:
    TicketStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]

    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 50 }
        password: { type: string, minLength: 1, maxLength: 128, format: password }

    CurrentUser:
      type: object
      required: [id, username, displayName, email]
      properties:
        id: { type: integer, format: int64 }
        username: { type: string }
        displayName: { type: string }
        email: { type: string }

    UserSummary:
      type: object
      required: [id, displayName]
      properties:
        id: { type: integer, format: int64 }
        displayName: { type: string }
        email: { type: string }

    CreateTicketRequest:
      type: object
      additionalProperties: false
      required: [title, description, priority, assigneeId]
      properties:
        title: { type: string, minLength: 3, maxLength: 150 }
        description: { type: string, minLength: 10, maxLength: 5000 }
        priority: { $ref: '#/components/schemas/Priority' }
        assigneeId: { type: integer, format: int64 }

    UpdateTicketRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title: { type: string, minLength: 3, maxLength: 150 }
        description: { type: string, minLength: 10, maxLength: 5000 }
        priority: { $ref: '#/components/schemas/Priority' }
        assigneeId: { type: integer, format: int64 }

    TransitionRequest:
      type: object
      additionalProperties: false
      required: [targetStatus]
      properties:
        targetStatus: { $ref: '#/components/schemas/TicketStatus' }

    CreateCommentRequest:
      type: object
      additionalProperties: false
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 2000 }

    TicketSummary:
      type: object
      required: [key, title, status, priority, assignee, reporter, updatedAt]
      properties:
        key: { type: string, example: TMS-42 }
        title: { type: string }
        status: { $ref: '#/components/schemas/TicketStatus' }
        priority: { $ref: '#/components/schemas/Priority' }
        assignee: { $ref: '#/components/schemas/UserSummary' }
        reporter: { $ref: '#/components/schemas/UserSummary' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    TicketPage:
      type: object
      required: [items, page, size, totalItems, totalPages]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TicketSummary' }
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer, format: int64 }
        totalPages: { type: integer }

    TicketSummaryCounts:
      type: object
      required: [assignedPending]
      properties:
        assignedPending:
          type: integer
          format: int64
          minimum: 0
          description: Tickets assigned to the signed-in user in status OPEN or IN_PROGRESS;
            equals `totalItems` of `GET /tickets?view=assigned` without q/status.

    TicketDetail:
      allOf:
        - $ref: '#/components/schemas/TicketSummary'
        - type: object
          required: [description, createdAt, version, allowedNextStatus, comments, history]
          properties:
            description: { type: string }
            resolvedAt: { type: string, format: date-time, nullable: true }
            closedAt: { type: string, format: date-time, nullable: true }
            version: { type: integer, format: int64 }
            allowedNextStatus:
              allOf: [{ $ref: '#/components/schemas/TicketStatus' }]
              nullable: true
              description: The only status this ticket may move to; null when CLOSED
            comments:
              type: array
              items: { $ref: '#/components/schemas/Comment' }
            history:
              type: array
              items: { $ref: '#/components/schemas/HistoryEntry' }

    Comment:
      type: object
      required: [id, author, body, createdAt]
      properties:
        id: { type: integer, format: int64 }
        author: { $ref: '#/components/schemas/UserSummary' }
        body: { type: string }
        createdAt: { type: string, format: date-time }

    HistoryEntry:
      type: object
      required: [id, actor, changeType, occurredAt]
      properties:
        id: { type: integer, format: int64 }
        actor: { $ref: '#/components/schemas/UserSummary' }
        changeType: { type: string, enum: [CREATED, FIELD_UPDATED, STATUS_CHANGED] }
        field: { type: string, nullable: true }
        oldValue: { type: string, nullable: true }
        newValue: { type: string, nullable: true }
        occurredAt: { type: string, format: date-time }

    FieldError:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }

    Problem:
      type: object
      description: RFC 9457 Problem Details with project extensions
      required: [title, status, code]
      properties:
        type: { type: string, default: about:blank }
        title: { type: string, description: Short user-friendly heading for the modal }
        status: { type: integer }
        detail: { type: string, description: User-friendly explanation }
        instance: { type: string }
        code:
          type: string
          enum:
            - VALIDATION_FAILED
            - INVALID_CREDENTIALS
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - TICKET_NOT_FOUND
            - INVALID_STATUS_TRANSITION
            - TICKET_CLOSED
            - VERSION_CONFLICT
            - PRECONDITION_REQUIRED
            - INTERNAL_ERROR
        errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        currentStatus: { $ref: '#/components/schemas/TicketStatus' }
        allowedNextStatus:
          allOf: [{ $ref: '#/components/schemas/TicketStatus' }]
          nullable: true
        traceId: { type: string }